Initial PNG Screenshot Layer
Replicated the original LunarG screenshot layer, but added support for the format we need to specify ranges of frames. Also updated vktrace to accept this format with a new command line option.

Currently, this still outputs PPM images.

User can now supply a range of frames: firstFrame-lastFrame,stepCount

examples:
   * To take a screenshot at every frame between frames 56-100
   * use "56-100,1"
   * If you only want every other frame between 56-100
   * use "56-100,2"

Note that if the 'lastFrame' doesn't fall on a 'stepCount' interval, a screenshot will NOT be taken there. "56-100,50" would only result in a screenshot at frame 56, because 56+50=106, and 106 is outside the 56-100 range.