vkreplay: Separate support for PngScreenshot command option vs ppm screenshot.
They are now two separate command line options:
-s or --Screenshot for ppm images, based on comma-separated list of
frames.
-png or --PngScreenshot for PNG images, based on frames identified by
<startFrame>-<endFrame>,<stepFrames>.

The associated layers need to be already enabled for the Vulkan loader
to use them, so the VK_INSTANCE_LAYERS and VK_DEVICE_LAYERS env vars
should be sent prior to running vktrace. The command line options are
just used to validate the input and set other variables for the layers
to know which frames to screenshot.