vktrace trim: Enable trimming to end of application.
Track the device queues to use when generating packets.

* Un-destroyed objects (particularly ShaderModules) are now destroyed by
 trim when the owning device is destroyed.

* We just need to have a valid handle to write into the trace packet,
so instead of actually calling into the device to get the queues we now
track a queue associated with the device and will use that handle.

* There is a possibility that a different queue is used to recreate the
object, but we don't yet have a case where this knowingly happens and
results in an artifact.

XCAP-536

Change-Id: I5637d9539dafd7845a59182a5d0a32fa653bf70c